Fix search, minor refactoring, add all not-added icons
When I was given the zip of the project to work on, there were many files not added to the project so I went ahead and added them.

I also added CocoaLumberjack for logging purposes, but I don't think there is any other logging in the project besides the lines I added for my debugging purposes and later removed.